Added userControllers Co-authored-by: Casey Walker casey.e.walker@gmail.com
